* master:
  Rlp decode returns Result (#8527)
  Node table sorting according to last contact data (#8541)
  Keep all enacted blocks notify in order (#8524)
  ethcore, rpc, machine: refactor block reward application and tracing (#8490)
  Consolidate crypto functionality in `ethcore-crypto`. (#8432)
  EIP 145: Bitwise shifting instructions in EVM (#8451)
  Remove expect (#8536)
  Don't panic in import_block if invalid rlp (#8522)
  Pass on storage keys tracing to handle the case when it is not modified (#8491)
  Fetching logs by hash in blockchain database (#8463)
  Transaction Pool improvements (#8470)
  More changes for Android (#8421)
  Enable WebAssembly and Byzantium for Ellaism (#8520)
  SecretStore: merge two types of errors into single one + Error::is_non_fatal (#8357)
  Hardware Wallet trait (#8071)
  Directly return None if tracing is disabled (#8504)
  Show imported messages for light client (#8517)

+++ miner/src/pool/queue.rs
@@ -282,11 +282,11 @@ impl TransactionQueue {
 		// We want to clear stale transactions from the queue as well.
 		// (Transactions that are occuping the queue for a long time without being included)
 		let stale_id = {
-			let current_id = self.insertion_id.load(atomic::Ordering::Relaxed) as u64;
+			let current_id = self.insertion_id.load(atomic::Ordering::Relaxed);
 			// wait at least for half of the queue to be replaced
 			let gap = self.pool.read().options().max_count / 2;
 			// but never less than 100 transactions
-			let gap = cmp::max(100, gap) as u64;
+			let gap = cmp::max(100, gap);
 
 			current_id.checked_sub(gap)
 		};
